// src/app/store/Client/Interface.ts
export interface ClientType {
  subscription_status?: string;
  first_name?: string;
  last_name?: string;
  email?: string;
  social_id?: any;
  files?: File;
  company_name?: string;
  id?: number;
  client_id?: number | string;
  role_id?: number;
  role?: string;
  country_code?: number | string;
  phone_number?: number | string;
  address?: string;
  user_image?: string;
  status?: string;
  deleted_at?: string;
  created_at?: string;
  updated_at?: string;
  userName?: string;
  agent_ids?: number[];
  address2?: string;
  city?: string;
  country?: string;
  zipcode?: number | string;
  state?: string;
  is_welcome_email?: number;
  last_subscription_detail?: any;
  next_subscription_detail?: any;
  is_edit_access?: any;
}

export interface filterType {
  start: number;
  limit: number;
  search: string;
  client_id?: number[] | string[];
  date?: any;
  currentDate?: any;
}

export interface deleteClientType {
  client_ids: number[];
}

export interface ProductDelete {
  product_id: number | string;
}

/**
 * The type definition for the initial state of the auth slice.
 */
export type initialStateProps = {
  status: string;
  subscriptionlist: any[];
  successMsg: string;
  errorMsg: string;
  fetchStatus?: string;
  getStatus?: string;
  clientDetail?: object;
  list: ClientType[];
  total_records: number;
  agentTotal_records: number;
  total_itemsSub: number;
  total_items: number;
  selectedColumn: string[];
  actionStatus: boolean;
  assignedAgentDetail: any[];
  assignedAgentDetailDashboard: any[];
  assignAccManagerDetail?: any[];
  toatalPage: number;
  currentPage?: number;
  managertotal_records?: number;
  total_itemsAcc?: number;
  actionStatusDisabled?: boolean;
  actionStatusClient?: boolean;
  is_welcome_email?: number;
  totalAgent?: number;
  agentListFetch?: boolean;
  resetActivity?: any[];
  clientResetActivity?: any[];
  dashBoardAgenda?: any[];
  fetchAgendaData?: boolean;
  agentActivity?: any[];
  globalfess?: any;
};
export interface ClientRootState {
  client: initialStateProps; // Add other slices if needed
}